Move Datasource console commands into App\DataSource\Console
