+++ app/Console/Kernel.php
@@ -16,9 +16,6 @@ class Kernel extends ConsoleKernel
         Commands\RoutesCommand::class,
         \Ushahidi\Console\Command\ConfigSet::class,
         \Ushahidi\Console\Command\ConfigGet::class,
-        \Ushahidi\Console\Command\DataSourceList::class,
-        \Ushahidi\Console\Command\DataSourceIncoming::class,
-        \Ushahidi\Console\Command\DataSourceOutgoing::class,
         \Ushahidi\Console\Command\Import::class,
         \Ushahidi\Console\Command\UserCreate::class,
         \Ushahidi\Console\Command\UserDelete::class,
