+++ src/App/DataSource/Console/IncomingCommand.php
@@ -9,7 +9,7 @@
  * @license    https://www.gnu.org/licenses/agpl-3.0.html GNU Affero General Public License Version 3 (AGPL3)
  */
 
-namespace Ushahidi\Console\Command;
+namespace Ushahidi\App\DataSource\Console;
 
 use Illuminate\Console\Command;
 
@@ -19,7 +19,7 @@ use Ushahidi\App\DataSource\DataSourceManager;
 use Ushahidi\App\DataSource\DataSourceStorage;
 use Ushahidi\App\DataSource\IncomingAPIDataSource;
 
-class DataSourceIncoming extends Command
+class IncomingCommand extends Command
 {
 
     /**
